refactor: auto-split test_property_based.py
